Add backend option to hide player's points
